// projects/common/src/lib/enums/utils/recaptcha-action.enum.ts
export enum RecaptchaActionEnum {
    LOGIN = 'LOGIN',
    SEND_NEW_PASSWORD = 'SEND_NEW_PASSWORD',
    CHANGE_PASSWORD = 'CHANGE_PASSWORD',
    EMAIL_PAYMENT = 'EMAIL_PAYMENT',
    CUSTOMER_LOGIN = 'CUSTOMER_LOGIN',
    CUSTOMER_REGISTER = 'CUSTOMER_REGISTER',
    CUSTOMER_SEND_NEW_PASSWORD = 'CUSTOMER_SEND_NEW_PASSWORD',
    CUSTOMER_CHANGE_PASSWORD = 'CUSTOMER_CHANGE_PASSWORD',
    CUSTOMER_USE_PAYMENT_METHOD = 'CUSTOMER_USE_PAYMENT_METHOD',
    CUSTOMER_COMPANY_INVITE = 'CUSTOMER_COMPANY_INVITE',
}